requirements.txt
Django==1.8.3
django-indexer==0.3.0
django-jinja==1.4.1
django-mail-queue==2.2.2
django-paging==0.2.5
django-simple-captcha==0.4.5
django-templatetag-sugar==1.0
djangorestframework==3.1.3
djoser==0.3.1
docutils==0.12
eventlet==0.17.4
greenlet==0.4.7
Jinja2==2.7.3
lockfile==0.10.2
MarkupSafe==0.23
Pillow==2.9.0
psycopg2==2.6.1
python-daemon==2.0.5
pytils==0.3
raven==5.5.0
six==1.9.0
